+++ app/admin/Players.js
@@ -1,0 +1,15 @@
+// ■ Usuarios registrados,
+// ■ Registro de uso de la plataforma por usuario
+// ■ Registro de participantes por etapas del fixture
+// ■ Permanencia en el sistema,
+// ■ Recordatorios enviados.
+'use client';
+import React from 'react'
+
+const Players = () => {
+  return (
+    <div>Players</div>
+  )
+}
+
+export default Players
