+++ app/admin/tournaments/Prizes.js
