+++ app/admin/tournaments/Results.jsx
