+++ app/admin/tournaments/[id].js
