+++ app/dashboard/page.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const page = () => {
+  return (
+    <div>holaaaaaaaaa</div>
+  )
+}
+
+export default page
