+++ app/layout.js
@@ -1,0 +1,12 @@
+export const metadata = {
+  title: 'Gambet'
+
+}
+ 
+export default function RootLayout({ children }) {
+ return (
+    <html lang="en">
+      <body>{children}</body>
+    </html>
+  )
+}
