+++ app/ranking/page.js
@@ -1,0 +1,20 @@
+"use client";
+
+import { Box } from "@mui/system";
+import React , {useRef} from "react";
+import RankingPodium from "./Podium";
+import RankingSearch from "./Search";
+
+const index = () => {
+  return (
+    <div>
+      <RankingPodium />
+      <Box sx={{display: "flex", justifyContent: "center",
+  alignItems: "center"}}>
+        <RankingSearch />
+      </Box>
+    </div>
+  );
+};
+
+export default index;
