--- src/pages/admin/Metrics.js
@@ -1,15 +1,0 @@
-// ■ Usuarios registrados,
-// ■ Registro de uso de la plataforma por usuario
-// ■ Registro de participantes por etapas del fixture
-// ■ Permanencia en el sistema,
-// ■ Recordatorios enviados.
-
-import React from 'react'
-
-const Metrics = () => {
-  return (
-    <div>Metrics</div>
-  )
-}
-
-export default Metrics
