--- src/pages/admin/Players.js
@@ -1,15 +1,0 @@
-// ■ Usuarios registrados,
-// ■ Registro de uso de la plataforma por usuario
-// ■ Registro de participantes por etapas del fixture
-// ■ Permanencia en el sistema,
-// ■ Recordatorios enviados.
-
-import React from 'react'
-
-const Players = () => {
-  return (
-    <div>Players</div>
-  )
-}
-
-export default Players
