--- src/pages/user/my-predictions.js
@@ -1,1 +1,0 @@
-
