--- src/pages/user/profile.js
@@ -1,1 +1,0 @@
-// ■ Formulario de perfil editable
