+++ yarn-error.log
@@ -1,0 +1,80 @@
+Arguments: 
+  /snap/node/7346/bin/node /usr/local/bin/yarn global dir --json
+
+PATH: 
+  /home/pablo/.local/bin/:/home/pablo/.local/bin:/home/pablo/.local/bin/:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:/usr/local/games:/snap/bin
+
+Yarn version: 
+  1.22.17
+
+Node version: 
+  16.20.0
+
+Platform: 
+  linux x64
+
+Trace: 
+  Error: EACCES: permission denied, write
+      at writeSync (node:fs:920:3)
+      at SyncWriteStream._write (node:internal/fs/sync_write_stream:26:3)
+      at writeOrBuffer (node:internal/streams/writable:391:12)
+      at _write (node:internal/streams/writable:332:10)
+      at SyncWriteStream.Writable.write (node:internal/streams/writable:336:10)
+      at JSONReporter._dump (/usr/local/lib/node_modules/yarn/lib/cli.js:49950:12)
+      at JSONReporter.log (/usr/local/lib/node_modules/yarn/lib/cli.js:49978:10)
+      at dir (/usr/local/lib/node_modules/yarn/lib/cli.js:36299:14)
+      at Object.<anonymous> (/usr/local/lib/node_modules/yarn/lib/cli.js:25983:25)
+      at Generator.next (<anonymous>)
+
+npm manifest: 
+  {
+    "name": "prode",
+    "version": "0.1.0",
+    "private": true,
+    "scripts": {
+      "dev": "next dev",
+      "build": "next build",
+      "start": "next start",
+      "lint": "next lint"
+    },
+    "dependencies": {
+      "@emotion/react": "^11.10.6",
+      "@emotion/styled": "^11.10.6",
+      "@mui/icons-material": "^5.11.11",
+      "@mui/lab": "^5.0.0-alpha.124",
+      "@mui/material": "^5.11.14",
+      "@mui/x-date-pickers": "^6.0.3",
+      "@reduxjs/toolkit": "^1.9.3",
+      "axios": "^1.3.4",
+      "date-fns": "^2.29.3",
+      "eslint": "8.36.0",
+      "eslint-config-next": "13.2.4",
+      "firebase": "^9.18.0",
+      "firebaseui": "^6.0.2",
+      "next": "13.2.4",
+      "next-redux-wrapper": "^8.1.0",
+      "react": "18.2.0",
+      "react-dom": "18.2.0",
+      "react-hot-toast": "^2.4.0",
+      "react-input-mask": "^2.0.4",
+      "react-intl": "^6.3.2",
+      "react-material-ui-carousel": "^3.4.2",
+      "react-multi-carousel": "^2.8.2",
+      "react-redux": "^8.0.5",
+      "react-slick": "^0.29.0",
+      "react-toastify": "^9.1.2",
+      "redux": "^4.2.1",
+      "redux-logger": "^3.0.6",
+      "slick-carousel": "^1.8.1",
+      "standard": "^17.0.0"
+    },
+    "devDependencies": {
+      "css-loader": "^6.7.3"
+    }
+  }
+
+yarn manifest: 
+  No manifest
+
+Lockfile: 
+  No lockfile
